86bz7xx5h migration from groovy to kotlin gradle

+++ app/src/main/java/study/me/please/ui/units/UnitsRepository.kt
@@ -12,6 +12,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import study.me.please.data.io.CollectionIO
 import study.me.please.data.io.FactIO
+import study.me.please.data.io.firebase.FirebaseCollections
 import study.me.please.data.io.subjects.ParagraphIO
 import study.me.please.data.io.subjects.UnitIO
 import study.me.please.data.room.CollectionDao
@@ -91,13 +92,22 @@ class UnitsRepository @Inject constructor(
         }
     }
 
+    /** updates a specific collection */
+    suspend fun updateCollection(collection: CollectionIO) {
+        return withContext(Dispatchers.IO) {
+            collectionDao.insertCollection(collection)
+        }
+    }
+
     private val cancellableScope = CoroutineScope(Dispatchers.IO)
 
     /** updates a specific unit within a collection */
     suspend fun updateFirebaseUnit(
         unit: UnitIO,
-        collectionUid: String
+        collectionUid: String?
     ) {
+        if(collectionUid == null) return
+
         cancellableScope.coroutineContext.cancelChildren()
         cancellableScope.launch {
             delay(NETWORK_UPDATE_DELAY)
@@ -111,8 +121,10 @@ class UnitsRepository @Inject constructor(
     /** updates a specific unit within a collection */
     suspend fun deleteFirebaseUnits(
         unitUidList: List<String>,
-        collectionUid: String
+        collectionUid: String?
     ) {
+        if(collectionUid == null) return
+
         val updateMap = mutableMapOf<String, Any>()
         unitUidList.forEach {
             updateMap["units.$it"] = FieldValue.delete()
@@ -139,10 +151,6 @@ class UnitsRepository @Inject constructor(
         }
     }
 
-    enum class FirebaseCollections {
-        COLLECTIONS
-    }
-
     /** Inserts or updates a new collection [collection] into the database */
     suspend fun insertCollection(collection: CollectionIO) {
         return withContext(Dispatchers.IO) {
@@ -151,7 +159,7 @@ class UnitsRepository @Inject constructor(
                     .collection(FirebaseCollections.COLLECTIONS.name)
                     .document(collection.uid)
                     .set(collection.apply {
-                        this.userUid = userUid
+                        this.authorUid = userUid
                     })
             }
